Created InstantSerializer.java for reusability
