+++ src/main/java/com/mindlinksoft/recruitment/mychat/InstantSerializer.java
@@ -1,0 +1,27 @@
+package com.mindlinksoft.recruitment.mychat;
+
+import java.lang.reflect.Type;
+import java.time.Instant;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonPrimitive;
+import com.google.gson.JsonSerializationContext;
+import com.google.gson.JsonSerializer;
+
+class InstantSerializer implements JsonSerializer<Instant> {
+    @Override
+    public JsonElement serialize(Instant instant, Type type, JsonSerializationContext jsonSerializationContext) {
+        return new JsonPrimitive(instant.getEpochSecond());
+    }
+    
+    public String createJson(Conversation conversation) {
+        GsonBuilder gsonBuilder = new GsonBuilder();
+        gsonBuilder.registerTypeAdapter(Instant.class, new InstantSerializer());
+
+        Gson g = gsonBuilder.create();
+
+        return g.toJson(conversation);
+    }
+}
